requirements.txt
pillow
pixoo
python-dateutil
pydantic
pydantic_settings
pyfiglet
requests
termcolor
